pyproject.toml
[tool.poetry]
name = "pytorch-training"
version = "0.1.0"
description = ""
authors = [" Sacha van Weeren <sacha.van.weeren@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.2"
matplotlib = "~=3.9"
hydra-core = "~=1.3.2"
pydantic = "~=1.10.7"
pytorch-lightning = "~=2.4"
tensorboard = "~=2.17"
torchvision = "~=0.19"
jupyterlab = "~=4.2"
#transformers = "^4"
datasets = "^2.15"
widgetsnbextension = "^4.0.13"
ipywidgets = "^8.1.5"
nodejs = "^0.1.1"
omegaconf = "^2.3.0"
scikit-learn = "^1.5.2"
boto3 = "^1.35.76"
tensorboardx = "^2.6.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"
types-decorator = "~=5.1"
types-redis = "~=4.6"
types-tabulate = "~=0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["cybulde"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.10"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0